✨ feat: add chaotic jump to pso

+++ evolutionary_programming/objective_function/base_function.pxd
@@ -7,4 +7,4 @@ np.import_array()
 
 cdef class BaseFunction:
     @classmethod
-    cpdef double evaluate(self, np.ndarray individual) except *
+    cpdef float evaluate(self, np.ndarray individual) except *
